+++ OpenDentBusiness/Data Interface/AppointmentRules.cs
@@ -10,6 +10,7 @@ namespace OpenDentBusiness{
 
 		///<summary></summary>
 		public static DataTable RefreshCache() {
+			//No need to check RemotingRole; Calls GetTableRemovelyIfNeeded().
 			string command="SELECT * FROM appointmentrule";
 			DataTable table=Cache.GetTableRemotelyIfNeeded(MethodBase.GetCurrentMethod(),command);
 			table.TableName="AppointmentRule";
@@ -18,6 +19,7 @@ namespace OpenDentBusiness{
 		}
 
 		private static void FillCache(DataTable table){
+			//No need to check RemotingRole; no call to db.
 			AppointmentRuleC.List=new AppointmentRule[table.Rows.Count];
 			for(int i=0;i<table.Rows.Count;i++) {
 				AppointmentRuleC.List[i]=new AppointmentRule();
@@ -31,6 +33,10 @@ namespace OpenDentBusiness{
 
 		///<summary></summary>
 		public static void Insert(AppointmentRule rule){
+			if(RemotingClient.RemotingRole==RemotingRole.ClientWeb) {
+				Meth.GetVoid(MethodBase.GetCurrentMethod(),rule);
+				return;
+			}
 			string command= "INSERT INTO appointmentrule (RuleDesc,CodeStart,CodeEnd,IsEnabled) VALUES("
 				+"'"+POut.PString(rule.RuleDesc)+"', "
 				+"'"+POut.PString(rule.CodeStart)+"', "
@@ -41,6 +47,10 @@ namespace OpenDentBusiness{
 
 		///<summary></summary>
 		public static void Update(AppointmentRule rule){
+			if(RemotingClient.RemotingRole==RemotingRole.ClientWeb) {
+				Meth.GetVoid(MethodBase.GetCurrentMethod(),rule);
+				return;
+			}
 			string command= "UPDATE appointmentrule SET " 
 				+ "RuleDesc = '"      +POut.PString(rule.RuleDesc)+"'"
 				+ ",CodeStart = '" +POut.PString(rule.CodeStart)+"'"
@@ -52,6 +62,10 @@ namespace OpenDentBusiness{
 
 		///<summary></summary>
 		public static void Delete(AppointmentRule rule){
+			if(RemotingClient.RemotingRole==RemotingRole.ClientWeb) {
+				Meth.GetVoid(MethodBase.GetCurrentMethod(),rule);
+				return;
+			}
 			string command="DELETE FROM appointmentrule" 
 				+" WHERE AppointmentRuleNum = "+POut.PInt(rule.AppointmentRuleNum);
  			Db.NonQ(command);
@@ -59,6 +73,7 @@ namespace OpenDentBusiness{
 
 		///<summary>Whenever an appointment is scheduled, the procedures which would be double booked are calculated.  In this method, those procedures are checked to see if the double booking should be blocked.  If double booking is indeed blocked, then a separate function will tell the user which category.</summary>
 		public static bool IsBlocked(ArrayList codes){
+			//No need to check RemotingRole; no call to db.
 			for(int j=0;j<codes.Count;j++){
 				for(int i=0;i<AppointmentRuleC.List.Length;i++){
 					if(!AppointmentRuleC.List[i].IsEnabled){
@@ -78,6 +93,7 @@ namespace OpenDentBusiness{
 
 		///<summary>Whenever an appointment is blocked from being double booked, this method will tell the user which category.</summary>
 		public static string GetBlockedDescription(ArrayList codes){
+			//No need to check RemotingRole; no call to db.
 			for(int j=0;j<codes.Count;j++) {
 				for(int i=0;i<AppointmentRuleC.List.Length;i++) {
 					if(!AppointmentRuleC.List[i].IsEnabled) {
